Gradio/requirements.txt
tensorflow==2.13
tensorflow_hub==0.14
scikit-learn
matplotlib
numpy
gradio
Pillow
